# Simplify `Routes`/`RoutePattern` trees