fixes #3408 
/claim #3408


**Benchmark Results**
```
Old:
[info] Benchmark                                   Mode  Cnt      Score       Error  Units
[info] RoutesBenchmark.benchmarkSmallDataZioApi   thrpt    3  32930.757 ± 20407.643  ops/s
[info] RoutesBenchmark.benchmarkSmallDataZioApi2  thrpt    3  42425.839 ±  7626.782  ops/s
[info] RoutesBenchmark.benchmarkSmallDataZioApi3  thrpt    3  51836.349 ±  2721.059  ops/s
[info] RoutesBenchmark.notFound1                  thrpt    3  22016.697 ±  2269.065  ops/s
[info] RoutesBenchmark.notFound2                  thrpt    3  19405.132 ±  2278.351  ops/s

[info] Benchmark                                            Mode  Cnt   Score    Error  Units
[info] RoundtripBenchmark.roundtripBenchmarkJavaClientAPI  thrpt    3  13.135 ± 10.142  ops/s

New:
[info] Benchmark                                   Mode  Cnt      Score      Error  Units
[info] RoutesBenchmark.benchmarkSmallDataZioApi   thrpt    3  39949.578 ±  612.926  ops/s
[info] RoutesBenchmark.benchmarkSmallDataZioApi2  thrpt    3  64732.621 ± 4345.785  ops/s
[info] RoutesBenchmark.benchmarkSmallDataZioApi3  thrpt    3  69094.709 ± 4062.645  ops/s
[info] RoutesBenchmark.notFound1                  thrpt    3  13494.307 ±  581.835  ops/s
[info] RoutesBenchmark.notFound2                  thrpt    3  21118.038 ±  423.684  ops/s

[info] Benchmark                                            Mode  Cnt   Score   Error  Units
[info] RoundtripBenchmark.roundtripBenchmarkJavaClientAPI  thrpt    3  16.851 ± 0.611  ops/s
```